% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{formatSea}
\alias{formatSea}
\title{Format Seahorse names}
\usage{
formatSea(seaList)
}
\arguments{
\item{seaList}{A character of Seahorse measurement names}
}
\description{
Function to give nice names of seahorse measurement
}
