% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{mscale}
\alias{mscale}
\title{Calculate modified row z-scores}
\usage{
mscale(x, center = TRUE, scale = TRUE)
}
\arguments{
\item{x}{The input matrix or datafram}

\item{center}{Centered by median}

\item{scale}{Scaled by 1.4826*MAD}
}
\description{
Row centered by median and scaled by median absolute deviation
}
