% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geneEnrichment.R
\name{plotEnrichmentBar}
\alias{plotEnrichmentBar}
\title{Barplot for enrichment analysis result}
\usage{
plotEnrichmentBar(resTab, pCut = 0.05, ifFDR = FALSE,
  setName = "Signatures")
}
\arguments{
\item{resTab}{A list object of result tables from runGSEA function}

\item{pCut}{The cutoff for p-values of signatures to be plot}

\item{ifFDR}{Whether to use FDR cutoff or raw p value cutoff (default)}

\item{setName}{The y axis label of the bar plot}
}
\description{
A function to plot enrichment analysis result from runGSEA function
}
