% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geneEnrichment.R
\name{runGSEA}
\alias{runGSEA}
\title{Function to run enrichment analysis in R}
\usage{
runGSEA(inputTab, gmtFile, GSAmethod = "gsea", nPerm = 1000)
}
\arguments{
\item{inputTab}{Ranked gene list for enrichment analysis, gene symbols as rownames and column stat contains the statistics for the ranking.}

\item{gmtFile}{A path to the gene signature file.}

\item{GSAmethod}{Method for enrichment analysis, currently supports GSEA and PAGE}

\item{nPerm}{Number of permutations for GSEA analysis}
}
\description{
A funtion to perform GSEA or PAGE analysis
}
