% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{sciPretty}
\alias{sciPretty}
\title{Create pretty scientific notation}
\usage{
sciPretty(n, digits = 2, bold = FALSE)
}
\arguments{
\item{n}{Number for formating}

\item{digits}{Digits of decimals to keep}
}
\description{
Function to generate pretty scientific notation format for plot label
}
